requirements.txt
click==7.0
matplotlib==3.0.3
numba==0.41.0
numpy==1.13.1
pandas==0.23.4
plotly==3.4.2
scipy==1.2.1
PyYAML==3.13
